requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
aiosqlite==0.19.0
openai==0.28.0
Pillow==10.0.0
pytesseract==0.3.10
